Update README.md
Officially finished with README.md file. The details to runserver are at the bottom as well as license detials "unlicensed". Folder paths and what each folder does is going to be redundant but this is to explain that we know what the app is doing and what is in each folder.